+++ microhapulator/pipe/reporter.py
@@ -67,6 +67,8 @@ class OverviewReporter:
             reads_are_paired=self.reads_are_paired,
             ambiguous_read_threshold=self.thresholds.ambiguous,
             read_length_threshold=self.thresholds.min_read_length,
+            discard_alert_threshold=self.thresholds.discard_alert,
+            gap_alert_threshold=self.thresholds.gap_alert,
         )
         return output